Implement the network activity indicator.
This needs to be done because Apple does not do any sort of fancy under-the-hood counting of the changes made to the networkactivityindicator property in UIApplication. Make a call to showNetworkActivityIndicator to ensure that the indicator is on. Only when all calls to showNetworkActivityIndicator have been matched with hideNetworkActivityIndicator will the indicator stop displaying.