Implement Playlist queueing.
Currently, playlists will be placed at the end of the second queue, so they won't get a chance to play until every song in the first queue and every song ahead of them in the second queue have finished playing. This behavior is likely to change but it's fine for now.