Fix behavior of SPTAudioStreamingController singleton in TrackPlayer.
While it's probably possible to reuse a SPTAudioStreamingController when playing different tracks, it's easier to just destroy the old one a create a brand new one. So now this isn't a singleton anymore, but as long as you only get a SPTAudioStreamingController by calling spotifyStreamingController then there should only ever be one instance of SPTAudioStreamingController running around.